Keep pending queries and their callbacks in an uservalue
This allows the context object to be garbage collected since it is not
anchored in the registry.

The resolve_async method now returns an uservalue instead of an integer
but the API does not change. Canceling a query works the same while not
needing a mapping of async_id to query object.